Added debug text, fixed multiple window
