+++ examples/3d_scene.rs
@@ -4,13 +4,11 @@ use bevy_mod_picking::*;
 fn main() {
     App::build()
         .add_resource(Msaa { samples: 4 })
-        .init_resource::<CursorEvents>()
         .add_default_plugins()
         .add_plugin(PickingPlugin)
         .add_plugin(DebugPickingPlugin)
         .add_startup_system(setup.system())
         .add_startup_system(set_highlight_params.system())
-        .add_system(get_picks.system())
         .run();
 }
 
@@ -69,32 +67,6 @@ fn setup(
             transform: Transform::from_translation(Vec3::new(4.0, 8.0, 4.0)),
             ..Default::default()
         });
-}
-
-pub struct CursorEvents {
-    cursor_event_reader: EventReader<CursorMoved>,
-}
-
-impl Default for CursorEvents {
-    fn default() -> Self {
-        CursorEvents {
-            cursor_event_reader: EventReader::default(),
-        }
-    }
-}
-
-fn get_picks(
-    pick_state: Res<PickState>,
-    mut cursor_events: ResMut<CursorEvents>,
-    cursor: Res<Events<CursorMoved>>,
-) {
-    match cursor_events.cursor_event_reader.latest(&cursor) {
-        Some(_) => println!(
-            "Top entity:\n{:#?}",
-            pick_state.top(PickingGroup::default())
-        ),
-        None => return,
-    };
 }
 
 fn set_highlight_params(mut highlight_params: ResMut<PickHighlightParams>) {
