+++ src/select.rs
@@ -37,9 +37,10 @@ pub fn select_mesh(
             selectable.selected = false;
         }
 
-        if let Some(pick_depth) = pick_state.top(PickingGroup::default()) {
-            if let Ok(mut top_mesh) = query.get_mut::<SelectablePickMesh>(pick_depth.entity) {
+        for (_group, pick) in pick_state.top_all() {
+            if let Ok(mut top_mesh) = query.get_mut::<SelectablePickMesh>(pick.entity) {
                 top_mesh.selected = true;
+                break;
             }
         }
     }
